requirements.txt
absl-py==2.3.1
accelerate==1.10.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.5
aiosignal==1.4.0
aistudio-sdk==0.3.7
alembic==1.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
attrs==25.3.0
audioread==3.0.1
autocorrect==2.6.1
av==13.0.0
bce-python-sdk==0.9.46
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
docker-pycreds==0.4.0
easyocr==1.7.2
einops==0.8.1
ffmpeg-python==0.2.0
filelock==3.19.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
gdown==5.2.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
imagesize==1.4.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.9
lazy_loader==0.4
Levenshtein==0.26.1
librosa==0.11.0
lightning-utilities==0.15.2
llvmlite==0.44.0
loguru==0.7.2
Mako==1.3.10
Markdown==3.9
MarkupSafe==3.0.2
matplotlib==3.10.6
modelscope==1.30.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
networkx==3.5
ninja==1.13.0
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai==1.102.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optuna==4.5.0
packaging==25.0
paddleocr==3.2.0
paddlepaddle==3.2.0
paddlex==3.2.1
pandas==2.2.3
pillow==11.1.0
pip==25.1
platformdirs==4.3.8
pooch==1.8.2
prettytable==3.16.0
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pypdfium2==4.30.0
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
RapidFuzz==3.13.0
regex==2025.7.34
requests==2.32.4
resampy==0.4.3
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safetensors==0.6.2
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.1
sentry-sdk==2.35.0
setproctitle==1.3.6
setuptools==78.1.1
shapely==2.1.1
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
SQLAlchemy==2.0.43
sympy==1.13.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.6.11
timm==1.0.19
tokenizers==0.21.4
torch==2.8.0+cu128
torchaudio==2.8.0+cu128
torchmetrics==1.4.2
torchvision==0.23.0+cu128
tqdm==4.65.2
transformers==4.49.0
triton==3.4.0
typing_extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
wandb==0.19.7
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
yarl==1.20.1
yt-dlp==2025.9.5
